Support for inkscape 1.0, fixes #1376
