+++ src/nl/hannahsten/texifyidea/util/LatexDistribution.kt
@@ -1,8 +1,6 @@
 package nl.hannahsten.texifyidea.util
 
 import nl.hannahsten.texifyidea.settings.TexifySettings
-import java.io.IOException
-import java.util.concurrent.TimeUnit
 
 
 /**
@@ -67,23 +65,7 @@ class LatexDistribution {
             return parsePdflatexOutput(runCommand("pdflatex", "--version"))
         }
 
-        private fun runCommand(vararg commands: String): String {
-            try {
-                val command = arrayListOf(*commands)
-                val proc = ProcessBuilder(command)
-                        .redirectOutput(ProcessBuilder.Redirect.PIPE)
-                        .redirectError(ProcessBuilder.Redirect.PIPE)
-                        .start()
 
-                // Timeout value
-                proc.waitFor(10, TimeUnit.SECONDS)
-                return proc.inputStream.bufferedReader().readText()
-            }
-            catch (e: IOException) {
-                e.printStackTrace()
-            }
-            return ""
-        }
 
         /**
          * Parse the output of pdflatex --version and return the distribution.
